Fix integration test reruns.
