+++ .gitignore
@@ -1,5 +1,6 @@
 # editor
 .vscode
+.idea
 
 # cache
 **/__pycache__
