+++ dags/csv_download_to_postgres/test_integration_csv_download_to_postgres.py
@@ -23,7 +23,7 @@ def test_dag_executes_with_no_errors(db_context: DB_Context):
     assert (
         execute_dag(
             "example_csv_to_postgres",
-            "2021-01-06",
+            "2021-01-01",
             {"TARGET_DB": credentials["database"]},
         )
         == 0
